backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi>=0.128.0",
    "httpx>=0.28.1",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=1.1.6",
    "langchain-text-splitters>=1.1.0",
    "pypdf>=5.9.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
]

[project.scripts]
company_search = "company_search:cli_main"
social_signal = "social_signals:cli_main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = ["main", "company_search", "social_signals"]

[tool.uv]
package = true
